requirements.txt
streamlit
tensorflow
keras
numpy
opencv-python
matplotlib
pillow
jupyter
scikit-learn
